requirements.txt
requests
beautifulsoup4
anthropic
streamlit
python-dotenv 